Create Building a party playlist
Building a party playlist
You're planning a party and want to build a playlist containing songs and the order in which they should be played.

You decide this is a great opportunity to build a list in Python! In this exercise, you will create your playlist variable and add the first three songs.

Instructions

Create a list variable called playlist, containing the following numeric and string values (in order):
1, "Blinding Lights", 2, "One Dance", 3, "Uptown Funk".
Print the list.