Create Subsetting lists 1
Subsetting lists
You've expanded your playlist. It still contains the song order and song names, but now you've also added the artist name for each song, as well as some additional songs!

It repeats the same order throughout the list - 1) song number, 2) song name, 3) artist name, repeat.

The new playlist has been printed in the IPython Shell for your reference.

This is a great opportunity for you to practice extracting specific information from your list.

Instructions 1/3

Find the name of the second song, which is the fifth element, in the playlist, and print the value.