Create Updating a variable with for loops
Updating a variable with for loops
You're planning to hold a party so you can show off your playlist to friends and family. You've hired a venue that has a maximum capacity of 30 people, and you'd like to keep track of tickets sold.

This is a perfect opportunity to use a for loop - it will increment the value of a variable each time a ticket is sold and then print a statement once the event is sold out.

Instructions

Create a variable called tickets_sold with a value of 0.
Create a variable called max_capacity with a value of 30.
Loop through a range starting at 1 and finishing at max_capacity plus one, naming the iterator as tickets.
Inside the for loop, add 1 to the value of tickets_sold.