Create Conditional looping with a dictionary
Conditional looping with a dictionary
Your friend at LLM Camp has provided you with a dictionary called courses. It contains the names of courses as keys and their associated topic as a value. There are three topics: "AI", "Programming", and "Data Engineering".

courses = {"LLM Concepts": "AI", 
           "Introduction to Data Pipelines": "Data Engineering", 
           "AI Ethics": "AI",
           "Introduction to dbt": "Data Engineering", 
           "Writing Efficient Python Code": "Programming",
           "Introduction to Docker": "Programming"}
This is a great chance to practice looping through dictionaries!

Instructions

Create a for loop to iterate over key, value in courses.items().
Check if the value is "AI", and print the key if so.
Else, check if the value is "Programming", printing the key if so.
Otherwise, print the key to confirm it is a "Data Engineering" course.