Create Conditional while loops
Conditional while loops
You've offered to help build a program that LLM Camp can use to promote an upcoming new course launching on the 26th of the month.

Today's date is the 22nd, and they have a pre-release version available from the 24th for all users who purchase a subscription on or before the 24th!

You'll need to build a custom workflow that provides tailored messages depending on the date.

Note that if your while loop takes too long to run, or your session is expiring, you might have created an infinite loop. In particular, remember to indent the contents of the loop using four spaces or auto-indentation, and make sure the conditions are such that the loop has a stopping point.

Instructions

Create a while loop based on current_date being less than or equal to release_date.
Check if current_date is less than or equal to 24 and, if so, print "Purchase before the 25th for early access"
Otherwise, check if current_date is equal to 25, printing "Coming soon!".
After all checks, increment current_date by one for each day that passes.